+++ tests/test_scoped_cleanups.py
@@ -4,11 +4,13 @@ import slash
 
 from .utils.suite_writer import Suite
 
+_MODULE_SCOPE_ADDER = 'slash.add_cleanup({0}, scope="module")'
+
 
 def test_module_scope(scoped_suite, file1_tests, file2_tests):
     file1_test = file1_tests[0]
     file1_end = file1_tests[-1].add_deferred_event(decorator='slash.add_cleanup')
-    file1_test_cleanup = file1_test.add_deferred_event(adder='slash.add_cleanup({0}, scope="module")')
+    file1_test_cleanup = file1_test.add_deferred_event(adder=_MODULE_SCOPE_ADDER)
 
     summary = scoped_suite.run()
     assert summary.events[file1_end].timestamp < summary.events[file1_test_cleanup].timestamp
@@ -25,6 +27,15 @@ def test_test_scoped_cleanups_in_session(checkpoint):
 
         assert not checkpoint.called
     assert checkpoint.called
+
+
+def test_errors_associated_with_correct_result(scoped_suite, file1_tests, file2_tests):
+    file1_test = file1_tests[0]
+    file1_test_cleanup = file1_test.add_deferred_event(adder=_MODULE_SCOPE_ADDER, extra_code=['assert 1 == 2'])
+    file1_test.expect_failure()
+
+    scoped_suite.run()
+
 
 @pytest.fixture
 def scoped_suite(suite, file1_tests, file2_tests):